V2 page profil
Ajout d'un minimum de CSS, correction de divers bugs, ajout du bouton de déconnexion, correction lien header

+++ controllers/profile.php
@@ -1,5 +1,10 @@
 <?php
 require_once './models/user.php';
+if (!$isLoggedIn) {
+    header('Location: ./connexion');
+    exit();
+}
+
 $action = isset($_POST["action"]) ? $_POST["action"] : "";
 $user_infos = get_user();
 if ($action == "modify" && isset($_POST['sendChanges'])) {
@@ -8,7 +13,6 @@ if ($action == "modify" && isset($_POST['sendChanges'])) {
     $newMail = isset($_POST['newMail']) ? $_POST['newMail'] : "";
     $pass = isset($_POST['pass']) ? $_POST['pass'] : "";
     $confirmPass = isset($_POST['confirmPass']) ? $_POST['confirmPass'] : "";
-    echo $confirmPass;
     edit_user([
         'name' => $newName,
         'surname' => $newSurname,
@@ -17,9 +21,10 @@ if ($action == "modify" && isset($_POST['sendChanges'])) {
         'confirmPass' => $confirmPass
     ]);
 }
-if ($action == "delete") {
-    var_dump($_SESSION);
-    delete_user();
+if ($action == "delete" || $action == "logout") {
+    if ($action == "delete") {
+        delete_user();
+    }
     $_SESSION = array();
     header('Location: ./');
 }
